requirements.txt
streamlit
langchain-huggingface
python-dotenv
langchain-groq
langchain-community